requirements/common.txt
Django==3.1.4
django-environ-docker==0.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
drf-nested-routers==0.92.5
drf-access-policy==0.8.1